Regions og fjernet fields vi ikke brugte
